Rename to InputRenderFactory.ts
